# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Bookstore API",
        "description": "Online Bookstore Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Cart List",
                "operationId": "22b703ec0cc7aca8223c2604ba590179",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "customer": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "test"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "test@gmail.com"
                                                            },
                                                            "role": {
                                                                "type": "string",
                                                                "example": "Admin"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "book": {
                                                        "properties": {
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Burning Heat"
                                                            },
                                                            "author": {
                                                                "type": "string",
                                                                "example": "Akiyoshi Rikako"
                                                            },
                                                            "publisher": {
                                                                "type": "string",
                                                                "example": "Elexmedia"
                                                            },
                                                            "year_released": {
                                                                "type": "string",
                                                                "example": "2021"
                                                            },
                                                            "genre": {
                                                                "type": "string",
                                                                "example": "thriller"
                                                            },
                                                            "price": {
                                                                "type": "integer",
                                                                "example": "120000"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "item": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "total": {
                                                        "type": "integer",
                                                        "example": 120000
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Cart Store",
                "operationId": "8ff5f58905eb254727c8bb30db32f873",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "book_id": {
                                                "type": "integer"
                                            },
                                            "qty": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "book_id": 1,
                                    "qty": 2
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "customer": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "test"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "test@gmail.com"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "Admin"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "book": {
                                                    "properties": {
                                                        "title": {
                                                            "type": "string",
                                                            "example": "Burning Heat"
                                                        },
                                                        "author": {
                                                            "type": "string",
                                                            "example": "Akiyoshi Rikako"
                                                        },
                                                        "publisher": {
                                                            "type": "string",
                                                            "example": "Elexmedia"
                                                        },
                                                        "year_released": {
                                                            "type": "string",
                                                            "example": "2021"
                                                        },
                                                        "genre": {
                                                            "type": "string",
                                                            "example": "thriller"
                                                        },
                                                        "price": {
                                                            "type": "integer",
                                                            "example": "120000"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "item": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 120000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}